Feature/dataseeder

+++ sentinel-backend/src/main/resources/db/migration/V5__schema_corrections.sql
@@ -1,8 +1,8 @@
 -- V5: Schema corrections — align DB with JPA entities
 
 -- Add coordinates to fact_incidents (IncidentEntity)
-ALTER TABLE fact_incidents ADD COLUMN latitude  DOUBLE;
-ALTER TABLE fact_incidents ADD COLUMN longitude DOUBLE;
+ALTER TABLE fact_incidents ADD COLUMN latitude  DOUBLE PRECISION;
+ALTER TABLE fact_incidents ADD COLUMN longitude DOUBLE PRECISION;
 
 -- Create fact_telemetry (TelemetryEntity)
 CREATE TABLE fact_telemetry (
@@ -10,9 +10,9 @@ CREATE TABLE fact_telemetry (
     timestamp            TIMESTAMP    NOT NULL,
     site                 VARCHAR(50)  NOT NULL,
     pipeline_section     VARCHAR(100),
-    pressure_psi         DOUBLE,
-    flow_rate_bph        DOUBLE,
-    temperature_celsius  DOUBLE,
+    pressure_psi         DOUBLE PRECISION,
+    flow_rate_bph        DOUBLE PRECISION,
+    temperature_celsius  DOUBLE PRECISION,
     valve_status         VARCHAR(30),
     sensor_id            VARCHAR(50)
 );
